+++ hgeometry/hgeometry.cabal
@@ -414,28 +414,29 @@ library
         HGeometry.VerticalRayShooting.PersistentSweep
 
     other-modules:
-       HGeometry.Polygon.Simple.Implementation
-       HGeometry.Polygon.Simple.InPolygon
+        HGeometry.HalfPlane.CommonIntersection.Chain
 
-       HGeometry.Polygon.Convex.Implementation
-       HGeometry.Polygon.Convex.MinkowskiSum
-       HGeometry.Polygon.Triangulation.Types
+        HGeometry.Polygon.Simple.Implementation
+        HGeometry.Polygon.Simple.InPolygon
 
-       HGeometry.PlaneGraph.Type
+        HGeometry.Polygon.Convex.Implementation
+        HGeometry.Polygon.Convex.MinkowskiSum
+        HGeometry.Polygon.Triangulation.Types
 
-       HGeometry.LineSegment.Intersection.Types
+        HGeometry.PlaneGraph.Type
 
-       HGeometry.LowerEnvelope.Type
-       HGeometry.LowerEnvelope.Connected.Type
-       HGeometry.LowerEnvelope.Connected.FromVertexForm
-       -- HGeometry.LowerEnvelope.Sample
+        HGeometry.LineSegment.Intersection.Types
+
+        HGeometry.LowerEnvelope.Type
+        HGeometry.LowerEnvelope.Connected.Type
+        HGeometry.LowerEnvelope.Connected.FromVertexForm
+        -- HGeometry.LowerEnvelope.Sample
 
 
-       -- HGeometry.LowerEnvelope.AtMostThree
-       -- HGeometry.LowerEnvelope.Triangulate
+        -- HGeometry.LowerEnvelope.AtMostThree
+        -- HGeometry.LowerEnvelope.Triangulate
 
-       HGeometry.VoronoiDiagram.ViaLowerEnvelope
-
+        HGeometry.VoronoiDiagram.ViaLowerEnvelope
